split nms.py and harris.py, create test feature
