+++ torchgeometry/feature/__init__.py
@@ -1,9 +1,9 @@
 from .harris import CornerHarris, corner_harris
-from .harris import NonMaximaSuppression2d, non_maxima_suppression2d
+from .nms import NonMaximaSuppression2d, non_maxima_suppression2d
 
 __all__ = [
-    "NonMaximaSupression2d",
     "non_maxima_suppression2d",
-    "CornerHarris",
     "corner_harris",
+    "NonMaximaSupression2d",
+    "CornerHarris",
 ]
